+++ NEWS.rst
@@ -5,4 +5,5 @@ fgallery 1.1: XX/XX/XXXX
 * Gallery generation speedup.
 * Improved handling of UTF-8/special file names.
 * Clicking on the main image shows the full image directly (if present).
+* Adapts to the current screen layout automatically.
 * Minor/cosmetic improvements.